Refactor origin_ui to web_ui - Reuse htpasswd auth and common APIs for all servers: /config /health - Move auth route to /api/v1.0/auth
