+++ docs/parameters.yaml
@@ -274,21 +274,6 @@ type: bool
 default: true
 components: ["origin"]
 ---
-name: Origin.UIPasswordFile
-description: >-
-  A filepath specifying where the Origin UI password file should be stored.
-type: string
-default: $ConfigBase/origin-ui-passwd
-components: ["origin"]
----
-name: Origin.UIActivationCodeFile
-description: >-
-  If the origin's web user interface has not yet been configured, this file will
-  contain the activation code necessary to turn it on.
-type: string
-default: $ConfigBase/origin-activation-code
-components: ["origin"]
----
 name: Origin.SelfTest
 description: >-
   A bool indicating whether the origin should perform self health checks.
@@ -520,6 +505,21 @@ description: >-
 type: string
 default: none
 components: ["origin", "director", "nsregistry"]
+---
+name: Server.UIActivationCodeFile
+description: >-
+  If the origin's web user interface has not yet been configured, this file will
+  contain the activation code necessary to turn it on.
+type: string
+default: $ConfigBase/server-web-activation-code
+components: ["origin", "cache", "nsregistry", "director"]
+---
+name: Server.UIPasswordFile
+description: >-
+  A filepath specifying where the Origin UI password file should be stored.
+type: string
+default: $ConfigBase/server-web-passwd
+components:  ["origin", "cache", "nsregistry", "director"]
 ---
