+++ origin_ui/src/helpers/login.tsx
@@ -1,5 +1,5 @@
 export async function isLoggedIn() {
-    let response = await fetch("/api/v1.0/origin-ui/whoami")
+    let response = await fetch("/api/v1.0/auth/whoami")
     if(!response.ok){
         return false
     }
