+++ web_ui/authorization.go
@@ -15,6 +15,7 @@
  * limitations under the License.
  *
  ***************************************************************/
+
 package web_ui
 
 import (
