+++ web_ui/ui.go
@@ -19,18 +19,46 @@
 package web_ui
 
 import (
+	"context"
 	"fmt"
+	"math/rand"
 	"net/http"
+	"os"
+	"os/signal"
+	"syscall"
 	"time"
 
 	"github.com/gin-gonic/gin"
 	"github.com/pelicanplatform/pelican/metrics"
 	"github.com/pelicanplatform/pelican/param"
+	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
 	ginprometheus "github.com/zsais/go-gin-prometheus"
+	"golang.org/x/term"
 )
 
-func ConfigureMetrics(engine *gin.Engine, isDirector bool) error {
+func getConfigValues(ctx *gin.Context) {
+	user := ctx.GetString("User")
+	if user == "" {
+		ctx.JSON(401, gin.H{"error": "Authentication required to visit this API"})
+		return
+	}
+	config, err := param.GetUnmarshaledConfig()
+	if err != nil {
+		ctx.JSON(500, gin.H{"error": "Failed to get the unmarshaled config"})
+		return
+	}
+
+	ctx.JSON(200, config)
+}
+
+func configureCommonEndpoints(engine *gin.Engine) error {
+	engine.GET("/api/v1.0/config", authHandler, getConfigValues)
+
+	return nil
+}
+
+func configureMetrics(engine *gin.Engine, isDirector bool) error {
 	// Add authorization to /metric endpoint
 	engine.Use(promMetricAuthHandler)
 
@@ -42,11 +70,96 @@ func ConfigureMetrics(engine *gin.Engine, isDirector bool) error {
 	prometheusMonitor := ginprometheus.NewPrometheus("gin")
 	prometheusMonitor.Use(engine)
 
-	engine.GET("/api/v1.0/health", func(ctx *gin.Context) {
+	engine.GET("/api/v1.0/health", authHandler, func(ctx *gin.Context) {
 		healthStatus := metrics.GetHealthStatus()
 		ctx.JSON(http.StatusOK, healthStatus)
 	})
 	return nil
+}
+
+func waitUntilLogin(ctx context.Context) error {
+	if authDB.Load() != nil {
+		return nil
+	}
+	sigs := make(chan os.Signal, 1)
+	signal.Notify(sigs, syscall.SIGINT, syscall.SIGTERM)
+
+	hostname := param.Server_Hostname.GetString()
+	port := param.Server_Port.GetInt()
+	isTTY := false
+	if term.IsTerminal(int(os.Stdout.Fd())) {
+		isTTY = true
+		fmt.Printf("\n\n\n\n")
+	}
+	activationFile := param.Server_UIActivationCodeFile.GetString()
+
+	defer func() {
+		if err := os.Remove(activationFile); err != nil {
+			log.Warningf("Failed to remove activation code file (%v): %v\n", activationFile, err)
+		}
+	}()
+	for {
+		previousCode.Store(currentCode.Load())
+		newCode := fmt.Sprintf("%06v", rand.Intn(1000000))
+		currentCode.Store(&newCode)
+		newCodeWithNewline := fmt.Sprintf("%v\n", newCode)
+		if err := os.WriteFile(activationFile, []byte(newCodeWithNewline), 0600); err != nil {
+			log.Errorf("Failed to write activation code to file (%v): %v\n", activationFile, err)
+		}
+
+		if isTTY {
+			fmt.Printf("\033[A\033[A\033[A\033[A")
+			fmt.Printf("\033[2K\n")
+			fmt.Printf("\033[2K\rPelican admin interface is not initialized\n\033[2KTo initialize, "+
+				"login at \033[1;34mhttps://%v:%v/view/initialization/code/\033[0m with the following code:\n",
+				hostname, port)
+			fmt.Printf("\033[2K\r\033[1;34m%v\033[0m\n", *currentCode.Load())
+		} else {
+			fmt.Printf("Pelican admin interface is not initialized\n To initialize, login at https://%v:%v/view/initialization/code/ with the following code:\n", hostname, port)
+			fmt.Println(*currentCode.Load())
+		}
+		start := time.Now()
+		for time.Since(start) < 30*time.Second {
+			select {
+			case <-sigs:
+				return errors.New("Process terminated...")
+			case <-ctx.Done():
+				return nil
+			default:
+				time.Sleep(100 * time.Millisecond)
+			}
+			if authDB.Load() != nil {
+				return nil
+			}
+		}
+	}
+}
+
+func ConfigureServerWebAPI(engine *gin.Engine, isDirector bool) error {
+	if err := configureAuthEndpoints(engine); err != nil {
+		return err
+	}
+	if err := configureCommonEndpoints(engine); err != nil {
+		return err
+	}
+	if err := configureMetrics(engine, isDirector); err != nil {
+		return err
+	}
+	// Redirect root to /view for web UI
+	engine.GET("/", func(c *gin.Context) {
+		c.Redirect(http.StatusFound, "/view/")
+	})
+	return nil
+}
+
+func InitServerWebUI() {
+	metrics.SetComponentHealthStatus(metrics.Server_WebUI, metrics.StatusWarning, "Authentication not initialized")
+
+	if err := waitUntilLogin(context.Background()); err != nil {
+		log.Errorln("Failure when waiting for web UI to be initialized:", err)
+		return
+	}
+	metrics.SetComponentHealthStatus(metrics.Server_WebUI, metrics.StatusOK, "")
 }
 
 func GetEngine() (*gin.Engine, error) {
