+++ web_ui/ui_windows.go
@@ -1,4 +1,5 @@
 //go:build windows
+
 /***************************************************************
  *
  * Copyright (C) 2023, Pelican Project, Morgridge Institute for Research
@@ -17,7 +18,7 @@
  *
  ***************************************************************/
 
-package origin_ui
+package web_ui
 
 import (
 	"github.com/pkg/errors"
